requirements.txt
RPi.GPIO
pigpio
pyserial
opencv-python
